Refactor lockbox_app.dart into separate screen files and rename to main.dart.
